solution type
