+++ src/core/sdpsos/solution.py
@@ -12,6 +12,7 @@ from ...utils import (
 
 
 class SolutionSDP(SolutionSimple):
+    method = 'SDPSOS'
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
